+++ guillotina_cms/api/__init__.py
@@ -1,4 +1,5 @@
 from . import components  # noqa
+from . import constraint  # noqa
 from . import login  # noqa
 from . import types  # noqa
 from . import workflow  # noqa