+++ guillotina_cms/interfaces/__init__.py
@@ -12,6 +12,7 @@ from .base import ICMSBehavior  # noqa
 from .content import IDocument  # noqa
 from .content import IImage  # noqa
 from .content import IFile  # noqa
+from .content import INews  # noqa
 from .events import IWorkflowChangedEvent  # noqa
 from .versioning import IVersioning  # noqa
 from .versioning import IVersioningMarker  # noqa