+++ guillotina_cms/permissions.py
@@ -5,12 +5,23 @@ configure.grant(
     role="guillotina.Manager")
 
 configure.permission('guillotina.ManageVersioning', 'Ability to modify versioning on an object')
+configure.permission('guillotina.ManageConstraints', 'Allow to check and change type constraints')
 
 configure.permission('guillotina.RequestReview', 'Request review permission')
 
 configure.grant(
     permission='guillotina.ManageVersioning',
     role='guillotina.Manager'
+)
+
+configure.grant(
+    permission='guillotina.ManageConstraints',
+    role='guillotina.Manager'
+)
+
+configure.grant(
+    permission='guillotina.ManageConstraints',
+    role='guillotina.ContainerAdmin'
 )
 
 configure.grant(